Bug fix
Fixes bug where if you dragged cards from the card selection window onto the trash it would reduce the count of that student type on the deck.

This caused the count of the card types to not be correct.

To fix I added a check to make sure the card being dropped on trash is from the deck otherwise it does not continue.